initial receiving code with input capture interrupts to decode transmitted signal
This code will detect a falling edge on the input capture pin of the
microcontroller and store the time at which the edge occured.  It will
then detect the next rising edge and take the difference between the
falling and rising edges to determine which signal has been sent and
which output LED to light up.  This code will be extended to receive
and decode 5 signals as follows:
1) left turn active
2) right turn active
3) brake signal active
4) left turn and brake signal active
5) right turn and brake signal active