receiving code with signal transmission begin interrupt and input capture interrupt to decode transmitted signal
This code will use a pin change interrupt to turn on input capture
interrupts and start a timer in the microcontroller to decode a
transmitted signal.  Using this method we will ensure that the
difference between the falling egde time and rising edge time (the
first and seconde parts of the coded signal respectively) is a positive
value.  When an initial change occurs on the input pin the Pin Change
Interrupt Vector will disable pin change interrupts and enable Input
Capture Interrupts to capture times of falling and then rising edges on
the input pin.  The difference between the falling and rising edge
times will determine which signal is being transmitted.