code to use pin change interrupts to turn 39KHz output wave on or off
This code uses Pin Change interrupts on pins 28,27, and 26 to turn a
39KHz 50% duty cycle square wave on or off.  When a pin change
interrupt occurs the Pin Change Interrupt Vector will determine which
pin caused the interrupt and turn a timer on in clear timer on compare
mode.  The timer will use output compare interrupts to generate a 39KHz
square wave on pin 23.  The output compare value is set at half the
period of the 39KHz wave , so at every output compare interrupt the
value on the output pin will be toggled high or low.